refactor(spf): rename provideTextTrackActors to setupTextTrackActors
Renames the behavior so it slots into the existing setup* family
(setupMediaSource, setupSourceBuffers) — all three create stateful
resources and publish them onto owners. The provide* prefix was the
only verb of its kind in playback/behaviors/dom/ and stood out as
inconsistent.

Also drops the orphaned .d.ts/.d.ts.map for the renamed module
(stale build artifacts that had been committed in error).

Touches:
- packages/spf/src/playback/behaviors/dom/{provide,setup}-text-track-actors.ts
- packages/spf/src/dom.ts (re-export)
- packages/spf/src/playback/engines/hls/engine.ts (import + local wrapper)
- packages/spf/src/playback/behaviors/dom/tests/load-text-track-cues.test.ts
- packages/spf/src/playback/behaviors/load-text-track-cues.ts (comment)

Public surface: @videojs/spf/dom now exports setupTextTrackActors
instead of provideTextTrackActors.
