+++ packages/spf/src/dom.ts
@@ -10,7 +10,7 @@ export { appendSegment } from './media/dom/mse/append-segment';
 export { flushBuffer } from './media/dom/mse/buffer-flusher';
 export { destroyVttResolver, resolveVttSegment } from './media/dom/text/resolve-vtt-segment';
 export { loadSegments } from './playback/behaviors/dom/load-segments';
-export { provideTextTrackActors } from './playback/behaviors/dom/provide-text-track-actors';
+export { setupTextTrackActors } from './playback/behaviors/dom/setup-text-track-actors';
 export type { CurrentTimeOwners, CurrentTimeState } from './playback/behaviors/dom/track-current-time';
 export { trackCurrentTime } from './playback/behaviors/dom/track-current-time';
 export type {
